# Conflicts:
#	nts.uk/uk.com/build.gradle
#	nts.uk/uk.com/shr/nts.uk.shr.com/src/main/java/nts/uk/shr/com/context/AppContexts.java

+++ nts.uk/uk.at/at.ctx/schedule/nts.uk.ctx.at.schedule.app/src/main/java/nts/uk/ctx/at/schedule/app/command/shift/estimate/guideline/CommonGuidelineSettingSaveCommandHandler.java
@@ -64,9 +64,13 @@ public class CommonGuidelineSettingSaveCommandHandler
 		} else {
 			this.commonGuidelineSettingRepo.update(commonGuidelineSetting);
 		}
+		// Find EstimateComparison
+		Optional<EstimateComparison> estCom = this.estimateComparisonRepo.findByCompanyId(companyId);
 		
-		// Delete EstimateComparison
-		this.estimateComparisonRepo.remove(companyId);
+		if (estCom.isPresent()) {
+			// Delete EstimateComparison
+			this.estimateComparisonRepo.remove(companyId);
+		}
 		
 		// Add EstimateComparison
 		this.estimateComparisonRepo.add(estComparison);
